Move utility functions to where they should be
